arg: Don't copy boxed types
In some places, an API passes a vfunc a boxed parameter and expects
the vfunc implementation to modify it. We can't copy boxed objects
in this case, as the modifications that the vfunc implementation did
won't apply to the copy.

We still need to not copy variants, as we ref them at the end of
the function.

https://bugzilla.gnome.org/show_bug.cgi?id=663492