Expose parent vfuncs on the parent prototype, prefixed with "do_"
As an example, the implementation of ClutterActor::get_preferred_width
will be exposed as Clutter.Actor.prototype.do_get_preferred_width.
We need this so that GObject subclasses can chain up when implementing
their own vfunc overrides.

The main big change here is that we now require the GType when using
gjs_define_function. This is because the implementation's GType may
not be the GICallableInfo's owner: ClutterRectangle has its own
get_preferred_width implementation, but since the vfunc isn't declared
in ClutterRectangleClass, introspection for ClutterRectangle doesn't
have a GIVFuncInfo. If a user created a ClutterRectangle subclass, in
order to chain up properly we need to expose
Clutter.Rectangle.prototype.do_get_preferred_width.

https://bugzilla.gnome.org/show_bug.cgi?id=663492