object: Fix handling classes without repository information
If a class doesn't have a GIObjectInfo, we should keep priv->info = NULL
NULL instead of picking that of a class, because this way we can
recognized it as resolve (this is the only case where we should look at
GType for methods).

Also, these classes should not be in an arbitrary namespace, instead
create an ad-hoc object inside imports.gi and stuff all of them there.
This is needed as soon custom objects will have dynamic JS classes.